Move importer scripts to bin/
