+++ bin/import_swb.py
@@ -5,11 +5,9 @@ from __future__ import absolute_import, division, print_function
 # This script needs to be run from the root of the DeepSpeech repository
 import sys
 import os
-sys.path.insert(1, os.path.join(sys.path[0], '..', '..'))
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
 
-import codecs
 import fnmatch
-import os
 import pandas
 import subprocess
 import unicodedata
@@ -18,41 +16,28 @@ import wave
 from util.text import validate_label
 
 def _download_and_preprocess_data(data_dir):
-    # Assume data_dir contains extracted LDC2004S13, LDC2004T19, LDC2005S13, LDC2005T19
+    data_dir = os.path.join(data_dir, "LDC97S62")
 
-    # Conditionally convert Fisher sph data to wav
-    _maybe_convert_wav(data_dir, "LDC2004S13", "fisher-2004-wav")
-    _maybe_convert_wav(data_dir, "LDC2005S13", "fisher-2005-wav")
+    # Conditionally convert swb sph data to wav
+    _maybe_convert_wav(data_dir, "swb1_d1", "swb1_d1-wav")
+    _maybe_convert_wav(data_dir, "swb1_d2", "swb1_d2-wav")
+    _maybe_convert_wav(data_dir, "swb1_d3", "swb1_d3-wav")
+    _maybe_convert_wav(data_dir, "swb1_d4", "swb1_d4-wav")
 
-    # Conditionally split Fisher wav data
-    _split_wav_and_sentences(data_dir,
-                             original_data="fisher-2004-wav",
-                             converted_data="fisher-2004-split-wav",
-                             trans_data=os.path.join("LDC2004T19", "fe_03_p1_tran", "data", "trans"))
-    _split_wav_and_sentences(data_dir,
-                             original_data="fisher-2005-wav",
-                             converted_data="fisher-2005-split-wav",
-                             trans_data=os.path.join("LDC2005T19", "fe_03_p2_tran", "data", "trans"))
-
-    # Conditionally split Fisher data into train/validation/test sets
-    train_2004, dev_2004, test_2004 = _split_sets(data_dir, "fisher-2004-split-wav", "fisher-2004-split-wav-sets")
-    train_2005, dev_2005, test_2005 = _split_sets(data_dir, "fisher-2005-split-wav", "fisher-2005-split-wav-sets")
+    # Conditionally split wav data
+    d1 = _maybe_split_wav_and_sentences(data_dir, "swb_ms98_transcriptions", "swb1_d1-wav", "swb1_d1-split-wav")
+    d2 = _maybe_split_wav_and_sentences(data_dir, "swb_ms98_transcriptions", "swb1_d2-wav", "swb1_d2-split-wav")
+    d3 = _maybe_split_wav_and_sentences(data_dir, "swb_ms98_transcriptions", "swb1_d3-wav", "swb1_d3-split-wav")
+    d4 = _maybe_split_wav_and_sentences(data_dir, "swb_ms98_transcriptions", "swb1_d4-wav", "swb1_d4-split-wav")
     
-    # The following file has an incorrect transcript that is much longer than
-    # the audio source. The result is that we end up with more labels than time
-    # slices, which breaks CTC. We fix this directly since it's a single occurrence
-    # in the entire corpus.
-    problematic_file = "fe_03_00265-33.53-33.81.wav"
-    train_2004.loc[train_2004.loc[train_2004["wav_filename"] == problematic_file], "transcript"] = "correct"
-
-    train_files = train_2004.append(train_2005)
-    dev_files = dev_2004.append(dev_2005)
-    dev_files = dev_2004.append(dev_2005)
+    swb_files = d1.append(d2).append(d3).append(d4)
+    
+    train_files, dev_files, test_files = _split_sets(swb_files)
 
     # Write sets to disk as CSV files
-    train_files.to_csv(os.path.join(data_dir, "fisher-train.csv"), index=False)
-    dev_files.to_csv(os.path.join(data_dir, "fisher-dev.csv"), index=False)
-    test_files.to_csv(os.path.join(data_dir, "fisher-test.csv"), index=False)
+    train_files.to_csv(os.path.join(data_dir, "swb-train.csv"), index=False)
+    dev_files.to_csv(os.path.join(data_dir, "swb-dev.csv"), index=False)
+    test_files.to_csv(os.path.join(data_dir, "swb-test.csv"), index=False)
 
 def _maybe_convert_wav(data_dir, original_data, converted_data):
     source_dir = os.path.join(data_dir, original_data)
@@ -69,9 +54,9 @@ def _maybe_convert_wav(data_dir, original_data, converted_data):
     # Loop over sph files in source_dir and convert each to 16-bit PCM wav
     for root, dirnames, filenames in os.walk(source_dir):
         for filename in fnmatch.filter(filenames, "*.sph"):
-            sph_file = os.path.join(root, filename)
-            for channel in ["1", "2"]:
-                wav_filename = os.path.splitext(os.path.basename(sph_file))[0] + "_c" + channel + ".wav"
+            for channel in ['1', '2']:
+                sph_file = os.path.join(root, filename)
+                wav_filename = os.path.splitext(os.path.basename(sph_file))[0] + "-" + channel + ".wav"
                 wav_file = os.path.join(target_dir, wav_filename)
                 print("converting {} to {}".format(sph_file, wav_file))
                 subprocess.check_call(["sph2pipe", "-c", channel, "-p", "-f", "rif", sph_file, wav_file])
@@ -80,14 +65,16 @@ def _parse_transcriptions(trans_file):
     segments = []
     with codecs.open(trans_file, "r", "utf-8") as fin:
         for line in fin:
-            if line.startswith("#") or len(line) <= 1:
+            if line.startswith("#")  or len(line) <= 1:
                 continue
 
             tokens = line.split()
-            start_time = float(tokens[0])
-            stop_time = float(tokens[1])
-            speaker = tokens[2]
-            transcript = " ".join(tokens[3:])
+            start_time = float(tokens[1])
+            stop_time = float(tokens[2])
+            transcript = validate_label(" ".join(tokens[3:]))
+
+            if transcript == None:
+                continue
 
             # We need to do the encode-decode dance here because encode
             # returns a bytes() object on Python 3, and text_to_char_array
@@ -99,58 +86,68 @@ def _parse_transcriptions(trans_file):
             segments.append({
                 "start_time": start_time,
                 "stop_time": stop_time,
-                "speaker": speaker,
                 "transcript": transcript,
             })
     return segments
 
-def _split_wav_and_sentences(data_dir, trans_data, original_data, converted_data):
+
+def _maybe_split_wav_and_sentences(data_dir, trans_data, original_data, converted_data):
     trans_dir = os.path.join(data_dir, trans_data)
     source_dir = os.path.join(data_dir, original_data)
     target_dir = os.path.join(data_dir, converted_data)
+    if os.path.exists(target_dir):
+        print("skipping maybe_split_wav")
+        return
+
+    os.makedirs(target_dir)
 
     files = []
 
     # Loop over transcription files and split corresponding wav
     for root, dirnames, filenames in os.walk(trans_dir):
-        for filename in fnmatch.filter(filenames, "*.txt"):
+        for filename in fnmatch.filter(filenames, "*.text"):
+            if "trans" not in filename:
+                continue
             trans_file = os.path.join(root, filename)
             segments = _parse_transcriptions(trans_file)
 
             # Open wav corresponding to transcription file
-            wav_filenames = [os.path.splitext(os.path.basename(trans_file))[0] + "_c" + channel + ".wav" for channel in ["1", "2"]]
-            wav_files = [os.path.join(source_dir, wav_filename) for wav_filename in wav_filenames]
+            channel = ("2","1")[(os.path.splitext(os.path.basename(trans_file))[0])[6] == 'A']
+            wav_filename = "sw0" + (os.path.splitext(os.path.basename(trans_file))[0])[2:6] + "-" + channel + ".wav"
+            wav_file = os.path.join(source_dir, wav_filename)
 
-            print("splitting {} according to {}".format(wav_files, trans_file))
+            print("splitting {} according to {}".format(wav_file, trans_file))
 
-            origAudios = [wave.open(wav_file, "r") for wav_file in wav_files]
+            if not os.path.exists(wav_file):
+                print("skipping. does not exist:" + wav_file)
+                continue
+
+            origAudio = wave.open(wav_file, "r")
 
             # Loop over segments and split wav_file for each segment
             for segment in segments:
                 # Create wav segment filename
                 start_time = segment["start_time"]
                 stop_time = segment["stop_time"]
-                new_wav_filename = os.path.splitext(os.path.basename(trans_file))[0] + "-" + str(start_time) + "-" + str(stop_time) + ".wav"
+                new_wav_filename = os.path.splitext(os.path.basename(trans_file))[0] + "-" + str(
+                    start_time) + "-" + str(stop_time) + ".wav"
                 new_wav_file = os.path.join(target_dir, new_wav_filename)
 
-                channel = 0 if segment["speaker"] == "A:" else 1
-                _split_wav(origAudios[channel], start_time, stop_time, new_wav_file)
+                _split_wav(origAudio, start_time, stop_time, new_wav_file)
 
                 new_wav_filesize = os.path.getsize(new_wav_file)
-                transcript = validate_label(segment["transcript"])
-                if transcript != None:
-                    files.append((new_wav_file, new_wav_filesize, transcript))
+                transcript = segment["transcript"]
+                files.append((new_wav_file, new_wave_filesize, transcript))
 
-            # Close origAudios
-            for origAudio in origAudios:
-                origAudio.close()
+            # Close origAudio
+            origAudio.close()
 
     return pandas.DataFrame(data=files, columns=["wav_filename", "wav_filesize", "transcript"])
 
 def _split_wav(origAudio, start_time, stop_time, new_wav_file):
     frameRate = origAudio.getframerate()
-    origAudio.setpos(int(start_time*frameRate))
-    chunkData = origAudio.readframes(int((stop_time - start_time)*frameRate))
+    origAudio.setpos(int(start_time * frameRate))
+    chunkData = origAudio.readframes(int((stop_time - start_time) * frameRate))
     chunkAudio = wave.open(new_wav_file, "w")
     chunkAudio.setnchannels(origAudio.getnchannels())
     chunkAudio.setsampwidth(origAudio.getsampwidth())
@@ -174,6 +171,16 @@ def _split_sets(filelist):
     return filelist[train_beg:train_end],
            filelist[dev_beg:dev_end],
            filelist[test_beg:test_end]
+
+def _read_data_set(filelist, thread_count, batch_size, numcep, numcontext, stride=1, offset=0, next_index=lambda i: i + 1, limit=0):
+    # Optionally apply dataset size limit
+    if limit > 0:
+        filelist = filelist.iloc[:limit]
+
+    filelist = filelist[offset::stride]
+
+    # Return DataSet
+    return DataSet(txt_files, thread_count, batch_size, numcep, numcontext, next_index=next_index)
 
 if __name__ == "__main__":
     _download_and_preprocess_data(sys.argv[1])
